Solve Challenge 4
Ugh...I spent a long time trying to improve/fix my scoring function, and was
confused at the garbage answers it was spitting out. I just assumed I was
missing the right answer. Dumb assumption; I forgot to decode the hex before
processing it.

Which, of course, is a point in favor of having a type specifically for RawBytes
that's returned by my encoding code, thus using type safety to avoid operating
on bytes which are not bytes.