Write test for challenge 5
Xor function already supported repeating-key since I knew it was going to come
up. This feels slightly like cheating but after missing the bug on challenge 4
last night and going in the entirely wrong direction to "fix" it I'm cool with
how easy this was. The test didn't even fail once!